Added conditions and validations for events and tickets. Implemented the passwordValidator, luhnAlgorithmCheck and sendCustomMail gems.
Added the conditions to pick only active and upcoming events and tickets. Added ordering of events, tickets and orders. Added extra validations in events and tickets for date and time. Implemented the passwordValidator, luhnAlgorithmCheck and sendCustomMail gems and used the same in code to validate password completixity, check if debit/credit card number is valid and send email on profile update and placing orders.